repos, services created

+++ OQM10P_HFT_2021221.Repository/Interfaces/IIssueRepo.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace OQM10P_HFT_2021221.Repository.Interfaces
 {
-    public interface IIssueRepo : IRepoBase<Issue, long>
+    public interface IIssueRepo : IRepoBase<Issue, int>
     {
     }
 }
